Update createRoomRequest method from RoomServiceImpl
A security check has been added, where the person creating the roomRequest must be the same as the one authenticated. This is achieved using the UnauthorizedActionException exception.

A limit has been set for sending roomRequest requests from a user to a room, the limit will be one. There cannot be more than one.